rpc_client.rs + JDsMempool
All the logic of the rpc_client is moved in a separate file
rpc_client.rs
Now the struct JDsMempool has also a client, which is inizialized
in the new() method. The get_mempool() method becomes update_mempool
This is far better as for updateing a mempool we do not reinizialise an
rpc_clint but we use every time the same.