new rpc client and non-blocking submit_solution
A new bitcoin rpc custom rpc client is added.
Before at the end of the trait method implementation
 handle_submit_solution in
jd-server/src/lib/job_declarator/message_handlers.rs
there was the part of the code responsible for submitting the
block to the bitcoin node via rpc. This was a blocker.
Now when submit_solution arrives, it is sent though a channel
to the mempool, that submits it to the node.
This task is managed asychronically in the main, so that we don't
have to wait that the rpc client submits the block to the
bitcoincore node.
The error management of JdsMempoolError enum is moved in mempool::error
module whenever possible in order to keep main clean.