Add MG test to replicate issue
Issue: when JDC changes upstream (because a SubmitShareError), it
disconnects also the TProxy.
To replicate this issue:
 - run a local template provider
 - run a pool applicationa that doesn't censor (i.e. the SRI pool)
 - run a jd-server application
 - from the MG test, run
`RUST_LOG=info cargo run
../../test/message-generator/test/change-upstream/change-upstream.json`
which is a mock for a pool that rejects the shares
 - launch a jd-client that connects to the pool's mock and the jds
 - launch the TProxy that connects to the jd-client
 - start mining with your favourite miner

At the first share relayed to the upstream the jdc disconnect the TProxy
All the config files needed are in `/test/config/change-upstream/`

The issue is fixed when this test is performed again, but the TProxy
reconnects to the jdc. The miner will likely continue mining on the
previous job (producing invalid shares) until a NewTemplate arrives